Task 8: modulation is here
