+++ index.html
@@ -232,8 +232,8 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
-  <!-- Подключение скрипта main.js -->
-  <script scr="./js/main.js"></script>
+  <!-- Подключение скрипта модуля main.js -->
+  <script type="module" src="./js/main.js"></script>
 
 </body>
 </html>
